Separate codec FFM bridge and make DecodedBatch a record
